# ngramoffset: add a dense asciiNgramOffset mapper