--- a/ngramoffset_test.go
+++ b/ngramoffset_test.go
@@ -16,6 +16,8 @@ package zoekt
 
 import (
 	"fmt"
+	"math/rand"
+	"sort"
 	"testing"
 )
 
@@ -55,6 +57,70 @@ func TestMakeArrayNgramOffset(t *testing.T) {
 	}
 }
 
+func TestMakeCombinedNgramOffset(t *testing.T) {
+	// The ascii / unicode ngram offset splitting is significantly
+	// more complicated. Exercise it with a more comprehensive test!
+	unicodeProbability := 0.2
+	ngramCount := 1000
+	ngramMap := map[ngram]bool{}
+
+	rng := rand.New(rand.NewSource(42))
+
+	randRune := func() rune {
+		if rng.Float64() < unicodeProbability {
+			return rune(0x100 + rand.Intn(0x80)) // Emoji
+		}
+		return rune('A' + rng.Intn('Z'-'A')) // A letter
+	}
+
+	for len(ngramMap) < ngramCount {
+		ngramMap[runesToNGram([3]rune{randRune(), randRune(), randRune()})] = true
+	}
+
+	ngrams := []ngram{}
+	for ng := range ngramMap {
+		ngrams = append(ngrams, ng)
+	}
+	sort.Slice(ngrams, func(i, j int) bool { return ngrams[i] < ngrams[j] })
+
+	offset := uint32(0)
+	offsets := []uint32{0}
+
+	for i := 0; i < len(ngrams); i++ {
+		// vary
+		offset += uint32(ngramAsciiMaxSectionLength/2 + rand.Intn(ngramAsciiMaxSectionLength))
+		offsets = append(offsets, offset)
+	}
+
+	m := makeCombinedNgramOffset(ngrams, offsets)
+
+	for i, ng := range ngrams {
+		want := simpleSection{offsets[i], offsets[i+1] - offsets[i]}
+		got := m.Get(ng)
+		if want != got {
+			t.Errorf("#%d: Get(%q) got %v, want %v", i, ng, got, want)
+		}
+		failn := ngram(uint64(ng - 1))
+		if getFail := m.Get(failn); !ngramMap[failn] && getFail != (simpleSection{}) {
+			t.Errorf("#%d: Get(%q) got %v, want zero", i, failn, getFail)
+		}
+		failn = ngram(uint64(ng + 1))
+		if getFail := m.Get(failn); !ngramMap[failn] && getFail != (simpleSection{}) {
+			t.Errorf("#%d: Get(%q) got %v, want zero", i, failn, getFail)
+		}
+	}
+
+	if t.Failed() || true {
+		t.Log(ngrams)
+		t.Log(offsets)
+		t.Log(m)
+	}
+}
+
+func (a combinedNgramOffset) String() string {
+	return fmt.Sprintf("combinedNgramOffset{\n  asc: %s,\n  uni: %s,\n}", a.asc, a.uni)
+}
+
 func (a *arrayNgramOffset) String() string {
 	o := "arrayNgramOffset{tops:{"
 	for i, p := range a.tops {
@@ -65,7 +131,7 @@ func (a *arrayNgramOffset) String() string {
 			// only one rune is represented here
 			o += fmt.Sprintf("%s: %d", string(rune(p.top>>10)), p.off)
 		} else {
-			o += fmt.Sprintf("%x: %d", p.top>>10, p.off)
+			o += fmt.Sprintf("0x%x: %d", p.top>>10, p.off)
 		}
 	}
 	o += "}, bots: {"
@@ -77,9 +143,24 @@ func (a *arrayNgramOffset) String() string {
 			// two ascii-ish runes (probably)
 			o += fmt.Sprintf("%s%s", string(rune(p>>21)), string(rune(p&runeMask)))
 		} else {
-			o += fmt.Sprintf("%x", p)
+			o += fmt.Sprintf("0x%x", p)
 		}
 	}
 	o += fmt.Sprintf("}, offsets: %v}", a.offsets)
 	return o
+}
+
+func (a *asciiNgramOffset) String() string {
+	o := "asciiNgramOffset{entries:{"
+	for i, e := range a.entries {
+		ng := ngramAsciiPackedToNgram(ngramAscii(uint32(e) >> 11))
+		length := e & ngramAsciiMaxSectionLength
+		if i > 0 {
+			o += ", "
+		}
+		o += fmt.Sprintf("%s: %d", ng, length)
+	}
+	o += fmt.Sprintf("}, chunkOffsets: %v}", a.chunkOffsets)
+	return o
+
 }