Most ngrams are of ASCII text and have small amounts of data. This
exploits this by packing three 7-bit runes and an 11-bit length into a
single 32-bit entry, with some extra logic to have periodic offsets to
easily reconstruct simpleSection{offset, length} with a fixed maximum
amount of extra computation.

This drops ngramOffset RAM from 3.6GB (69% of total) to 2.3GB (59%),
with asciiNgramOffsets taking 1.7GB (44%) and arrayNgramOffsets taking
0.6GB (15%).

Unfortunately, this increases CPU usage when loading the index--
readIndexData increases from 13s to 18s, with the main call of
readNgrams cpu time doubling from 5s to 10s.

This also adds a more comprehensive test to exercise the various
boundary conditions of the ascii/unicode splitting logic.

---

\>95% of running zoekt-webserver memory usage is due to indexData structs, so this will effectively shrink the steady-state memory usage of zoekt servers hosting many shards.

These memory optimizations have two main methods:

1) lazily decode deltas during search time when possible (filenameNgrams, runeDocSections). These attributes are only used for somewhat uncommon search types `file:` and `symbol:`, and the volume of decoded deltas is still _far_ less than the amount of deltas that are decoded during a normal indexed search operation, so the impact should be negligible.

2) convert some excessively large indexes with `O(1)` access times to smaller `O(log(n))` access time indexes (ngrams, runeOffsets). This is the motivation for the patch series, where the overhead of `map[ngram]simpleSection` (a 64 bit to 64 bit map) could be greatly reduced. The slightly increased time complexity _shouldn't_ matter, given that most search CPU time is spent processing the actual trigram lists.

As a bonus, index loading is 2.2x faster, going from 5.6s->2.5s wall clock and 54s->22s cpu time.

In terms of memory usage, loading 18855 shards from popular repositories (all the shards on one node of sourcegraph.com) went from this:
![image](https://user-images.githubusercontent.com/211868/122507163-ef44d900-cfbc-11eb-8051-50b5e4104649.png)

To this:
![image](https://user-images.githubusercontent.com/211868/122506983-9c6b2180-cfbc-11eb-8a3d-bc944cceab47.png)