@rmmh 2021-06-18
Split into #92 #93 #94. Leaving this open for future rebasing and talking about asciiNgramOffset.