@rmmh 2021-07-28
I was clicking "resolved" as a shorthand for "done". I'll be more verbose.